requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
black==22.3.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.4
discord.py==1.7.3
idna==3.3
idna-ssl==1.1.0
importlib-metadata==4.8.3
multidict==5.2.0
mypy-extensions==0.4.3
oauthlib==3.2.0
pathspec==0.9.0
platformdirs==2.4.0
python-dateutil==2.8.2
python-dotenv==0.20.0
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
tomli==1.2.3
typed-ast==1.5.3
typing-extensions==4.1.1
urllib3==1.26.9
yarl==1.7.2
zipp==3.6.0
